Attempt to wrap cython around the C++ leveldb API
I just noticed that the leveldb library has a C++ class API. It would make more sense to use this rather than the C API.
In theory we should be able to statically link the leveldb library and compile everything at the same time.
This should fix the issues with installing on different platforms.