Added locking
The methods that get and modify data may be run in parallel but they may not be run in parallel with the closing logic.
The closing logic cannot be run in parallel with itself.
This implements this behaviour using shared and unique locks from C++
The close method can now be called multiple times. Only the first call will actually do anything.